Divide user balance

+++ TeleportApi/database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,8 +22,8 @@ class CreateUsersTable extends Migration
             $table->string('email')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password')->nullable();
-            $table->decimal('balance_contractor')->nullable();
-            $table->decimal('balance_employer')->nullable();
+            $table->integer('balance_contractor')->nullable();
+            $table->integer('balance_employer')->nullable();
             $table->string('employer_tariff')->nullable();
             $table->string('contractor_tariff')->nullable();
             $table->rememberToken();
